rename async generate stream file
